Limits local witness path searches in edge-based CH.
* when a certain (right now hard-coded) number of original edges have
been settled in the witness path search, only edges are settled that might
represent the original shortest path (current incoming edge + loops
+ current outgoing edge). this is needed because the shortest connection
between the neighbor nodes has to be found as well and we can not simply
stop the whole witness path search when a certain number of edges is
settled or a certain weight is exceeded.

* performance is still bad, but much better than previously:
Bremen map: preparation time ~45s, query time ~5ms.

Signed-off-by: ammagamma <contactammagamma@gmail.com>