Adapt FootFlagEncoder to new GH interface (almost done)
FootFlagEncoder has been adapted to use the new EncodedValues.
The remaining issue is the overriding of method getSpeed
which is package-private in GH and, therefore, not visible in ORS.
This intermediate state is committed in order to enable to proceed
with other things until we found a solution for the above problem.