Check for null pointers in shortcut expansion (#563)
This stops the system from crashing during the core alt build process when a null pointer is hit. 

This is an adaptation of the Path4CH expansion algo. The original algo does not properly apply to our core and therefore produces NPEs sometimes. For now this is fixed by a brute force algo that just tries to expand all edges and checks for null every time

Authored by @HendrikLeuschner 

* Check for null pointers in shortcut expansion

* adapt signature

* add todo message to fix