Improve Travis build process
Resolve maven dependencies in a separate step before actually building
and deploying the ORS web server for the API-tests.

Reason for this change: We had a long artificial pause after starting the
ORS web server in order to compensate for varying download times in case
of high network traffic. This waiting time is now reduced to the duration
of graph building plus a safety buffer.