Correctly initialize EdgeExplorer
As stated in the graphhopper documentation, only one instance of EdgeExplorer
should ever be created. This instance is then transferred to the
CentralityAlgorithm, since it is a central part of its workings.

The EdgeExplorer used is set up to only accept outgoing edges valid for
the flagEncoder derived from the given profile.