remove todo for testing invalid excludenodes
Informing the user about a invalid excludeNode would require a
significant overhead in the calculation of nodes in the bounding box.
It could not inform the user about removing nodes he doesn't want to
remove. This has to be handled correctly by the user and is not handled
by the centrality endpoint.