Enable logging in AppConfig.java
AppConfig gets called first during initializing the logger itself.
It was therefore not possible to load the default configuration from DEFAULT_LOGGING.json,
add an Appender, or change the layout of the current default sysOut appender of
the root Logger.

The current solution is to adjust the log level to INFO and reverting
to previous log level afterwards.