`sett` as default minimum wheelchair surface type
The osm wiki denotes the previous default of `cobblestone:flattened` as
not to use. All three of `sett`, `unhewn_cobblestone` and
`cobblestone:flattened` are encoded as the same value, so no changes to
the codebase are required, it is only a matter of correct documentation.

Since `sett` is way more prevalent, it will be used as the default.