Access vehicle restriction values in hgv storage as bytes rather than shorts
This addresses a possible memory alignment issue caused by the fact that hgv storage entries are 11 bytes long. As this lenght is not a multiple of 2 it can happen that the two bytes constituting a short value get distributed over two different byte array segments. However, unlike the byte-reading/writing methods the methods to directly access shorts do not allow them to be split between segments.