Proper handling of virtual edges in RPHAST downward search
Use the edge iterator state extracted from the subgraph edge iterator for computing edge weight during RPHAST downwards search in order to be able to resolve original edge IDs in case of virtual edges. This is necessary e.g. for the HgvAccessWeighting to be able to reliably query the HGV storage.