fix: rename csvColumn API parameter variable
Swagger seems to generate API parameters from getter and setter names.
Corresponding to the csv_column API parameter, getters and setters are
called getCsvColumn and setCsvColumn. This generates an API parameter
called csvColumn, that isn't there.

Renaming the variable "links" the variable and getters/setters, and the
csvColumn parameter vanishes.