build(spring-boot): remove fork setting
The fork setting set to false mixes up run paths but allows for local debug session. When set to true it disables the ability for a local debug session but has the expected run paths. To have local debug sessions one can use  -Dspring-boot.run.fork=false to disable forking when needed. This is only possible when the fork settings is not set at all in the pom.