chore(cleanup): Remove unused custom org.json and the jsonUtility classes
Some of the package classes that were added a few years ago are not even used. They can be removed. Additionally, some of the classes are so small and generic, so that they didn't change since 2017. Besides that the main json classes provide all necessary functionality to work. Only jsonUtility was using the ols classes but itself is never used anywhere.