feat(ors-config): add CORS settings
there have been CORS issues when using our map clients with local
ors instances during development due to Problems with what headers are
accepted and exposed by the backend.
With the spring-boot update and the new run configurations using spring,
setting these options properly is now possible for local instances using
`spring-boot:run` or mvn version of it.

However, the docker approach is still using catalina + tomcat where
the settings would be passed differently, so they are not respected in
a local backend running with docker.

- add addCorsMappings function to ApiConfig class
- extent ors-config-sample.json with default cors settings working with
  all origins and accepting `Authorization` headers.
- if the api settings are missing from the ors config file,
 default values are used that correspond with the entries
 in the ors-config-sample.json