tests(github,action): Add CORS GitHub action tests
These tests check the tomcat + war file deployment used in Docker,
which cannot be checked via the integration or unit tests.
The tests check with the wildcard from the default config first
and then replace the wildcard with https://example.org.
With that set, it checks for correct CORS issues by requesting
with origin https://example.com.