ci(docker): Fix the docker build github workflow
Currently, the workflow doesn't seem to correctly build for arm64 platforms. This is fixed now for the nightly images and the release workflow. arm64 tends to be really slow on building graphs while being emulated on an amd64 machine. Because of this the arm64 docker tests are only executed when the PR is set to active. Additionally, the grype checks now test both amd64 and arm64 platforms correctly.