chore: add GitHub issue form templates
the previous issue template was outdated and people often had
trouble with the markdown format.

By using issue forms, formatting is less error prone,
information more structured and requirements more clear.
The forms can be tailored to the issue category and labels
(or even assignees) automatically applied.

Templates were added for the following categories:
- bug reports: thorough information required
- docs: issue with the GitHub pages documentation
- feature requests
- other issues: simple form for quick issues or technical
problems not fitting the bug report

The config.yml allows for external links which point to the
forum for questions and the ors-map-client repo for issues
with the map application.

Co-authored-by: Julian Psotta <22392104+MichaelsJP@users.noreply.github.com>
Co-authored-by: Jakob Schnell <Jakob.Schnell@heigit.org>