refactor(config): remove MatrixServiceSettings from matrix algorithms
This commit removes the usage of the static MatrixServiceSettings
from AbstractManyToManyMatrixAlgorithm in favor of the new
configuration system. Unfortunately, the current structure of
the matrix algorithms forced this commit to pass around the
boolean "hasInfiniteUTurnCosts" through several classes. This
should be cleaned up together with the matrix algorithms in
a later commit.