fix: re-introduce edge metrics cache in matrix path extractor
Matrix path extractor uses a hashmap-based cache to keep track of encountered edges. This is a performance improvement in order to avoid re-computing metrics for edges which have been previously visited. The original solution has been removed in cbecd7d1fa8022e61ae3a0495ed2437550cc543e because it had a flaw of storing the cache entries under non-unique keys computed as sum of edge id and adjacent node id. The caching is reintroduced with a fix of using unique edge keys.