style(Java14): replace switch with new switch types
replaces normal switch with enhanced switch statements or expressions.

Note: we use a redundant default case where the
default case was the same as one of the normal cases.
This is to avoid a Language Level change to 17 (Preview),
which is only available with the --enable-preview compile
command and would require a more complex java setup.

example:
new switch expression requiring the pattern matching
    case Instruction.CONTINUE_ON_STREET, default -> InstructionType.CONTINUE;

redundant default:
    case Instruction.CONTINUE_ON_STREET -> InstructionType.CONTINUE;
    default -> InstructionType.CONTINUE;

This could be avoided if the default case actually does
something different or the duplicate normal case removed.

As this touches implementation details, it is skipped for now.