feat(config): Add bash script to rewrite the application.yml to the ors-config.yml
This feature intents to keep the application.yml and the ors-config.yml in sync. The ors-config.yml is intented as a minimal working configuration for users that want an up and running system without configuration. The in-depth configs are still in the ors-config.yml but commented out. The sync script will make sure in a workflow that if the original application.yml changes, the ors-config.yml mirrors the new changes but commented out.