fix(docker-compose): Adjust the image tag to local/openrouteservice:latest
This ensures that users that checkout the compose from main will either fail or if it was cloned with the repo will have the latest repo stage build as their docker image. The image line is adjusted in the release action so that the release artifacts compose files will always point to their subsequent image version on docker hub.