fix(config): Resolve absolute paths for the config
The config is handling paths internally without modifying it (except once for graphs) and expects the relative rundir not to change. The consistency of the rundir is not given so the paths should be resolved to absolute paths when the engine starts.
This is missing the paths from the external storages that are in graphhopper.