feat: configure graph management explicitly
Changes on main, that were merged into the branch, broke the implementation of the detection if local graph building or a graph repo should be used. This approach was controversial anyway. We replaced it with explicit properties for graph management. This has the side effect, that a graph is built locally, if it is not found in the repository and an existing source_file is configured. Whether this behavior is desirable is questionable, because the configured pbf could be of a different extent than configured.