fix: Rework raw user settings hierarchy
These changes allow that all settings are null in the beginning and only the raw settings from the users are initialized. This allows to later differentiate what settings have been given by the user and which not.
What is missing is the remaining logic: Set the default_config as the next hierarchy and after all the remaining defaults, that are necessary.