perf: Streamline the container creation
From now on the test classes extend the abstract container initializer. As a consequence they all run the @BeforeAll from the initializer. The way they are configured results in a parallel execution that goes class after class. Only the methods and parametrized executions in a test class are executed in parallel per class. That being said, layers are cached and graphs are build centrally before any test is executed.