CI(integration-tests): Add the integration-tests.yml workflow
The new workflow first builds and caches the builder image. This image is big and clunky and would only increase the parallel test time massively. The testing graph is build in the same way. The image ID from the image is used to check compatibility with the graph. When all that works the integration tests are started in parallel for all scenarios and test classes.