feat: enable `way_surface` and `way_type` encoded values through config file
For the sake of backwards compatibility, for now encoded values are set based on the presence of a corresponding config entry in `ext_storages`. This approach is only a temporary solution in order to avoid changes to the config file structure which would probably require a major version bump. This however, would only make sense once the rest of extended storages has been converted to encoded values.