ci: add composite actions
Our workflows are very redundant in many places. It is hard to maintain it. This will abstract the most common code redundancies and have them in one maintainable place. This allows for central version changes and functionality testing.

The following actions are implemented. They will be used in later PRs:

setup-build-environment is the absolute basic environment setup for all workflows. It can set up all necessary environments with simple flags. It also streamlines the actions versions and docker caches.

parse-release-tag is now well encapsulated and testable release tag parsing function.